preliminary retriever with context
